backend/pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = ""
authors = [
    {name = "thallys",email = "thallys.rc@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)"
]

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
